unify downsample helpers
